Refactor logging
Moved log and Message to a util class and updated the overwrite logic to allow consecutive lines to be overwritten.
